move PerformancePrinter
