+++ proteus-compiler-kotlin/src/main/kotlin/lang/proteus/api/ProteusCompiler.kt
@@ -1,45 +1,11 @@
 package lang.proteus.api
 
 import lang.proteus.api.performance.CompilationPerformance
-import lang.proteus.api.performance.ComputationTime
 import lang.proteus.api.performance.ComputationTimeStopper
+import lang.proteus.api.performance.PerformancePrinter
 import lang.proteus.printing.ConsolePrinter
 import lang.proteus.printing.PrinterColor
 import lang.proteus.syntax.parser.SyntaxTree
-
-class PerformancePrinter {
-
-    private val consolePrinter = ConsolePrinter()
-    fun print(performance: CompilationPerformance) {
-        printComputationTime("Lexing", performance.lexingTime)
-        printComputationTime("Parsing", performance.parsingTime)
-        printComputationTime("Evaluation", performance.evaluationTime)
-    }
-
-    private fun printComputationTime(prefix: String, time: ComputationTime) {
-        consolePrinter.setColor(PrinterColor.WHITE)
-        consolePrinter.print("$prefix: ")
-        consolePrinter.setColor(getColorFromTime(time.time))
-        consolePrinter.print(formatTime(time.time))
-        consolePrinter.println()
-    }
-
-    private fun getColorFromTime(time: Long): PrinterColor {
-        return when {
-            time < 10 -> PrinterColor.GREEN
-            time < 100 -> PrinterColor.YELLOW
-            else -> PrinterColor.RED
-        }
-    }
-
-    private fun formatTime(time: Long): String {
-        return if (time < 1000) {
-            "${time}ms"
-        } else {
-            "${time / 1000}s"
-        }
-    }
-}
 
 class ProteusCompiler(private var variables: Map<String, Any>) {
     fun compile(line: String, verbose: Boolean = false): CompilationResult {
